ARM: config: aspeed-g5: Add DW UART and ADT7475
This adds the Synopsis DesignWare UART quirks, which is used by the
AST2600 A0 to workaround an issue.

The ADT7475 is an I2C device present on the AST2600 EVB and is useful
for testing that I2C works.

OpenBMC userspace does not yet understand the "new" fsi device node
layout, so disable it for now.

OpenBMC-Staging-Count: 1
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>