fsi: aspeed: Implement byte and half word writes
The driver did not previously support non-word size transfers. The
fsi-i2c driver attempts accesses of this size, so we require it now.

This creates three functions for doing word, half word and byte
writes. This was done to avoid having to decide which length the
transfer was twice: both in aspeed_master_write to determine how much of
the void * to copy, and again in opb_write to look up the opcode.

It we could have passed the opcode in aspeed_master_write, but this
allows the majority of opb_writes in this driver to omit the opcode/size
parameter, as most writes are full words. It also means call sites get
the types checked.

OpenBMC-Staging-Count: 1
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Acked-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Joel Stanley <joel@jms.id.au>