fsi: aspeed: Clean up defines and documentation
Some of the registers lacked documetation but now can be properly named.

Rename FW to FULLWORD so no one thinks it means 'firmware'.

The "enable DMA" comment was setting the FSI access direction for DMA.
As the driver does not yet use DMA it can be removed without effect.

OpenBMC-Staging-Count: 1
Acked-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>