tpm: tpm_tis: Add check_data handle to tpm_tis_phy_ops
In order to validate data integrity we need to compute the crc over the data
sent in lower layer (I2C for instance).
To do that tpm_tis_check_data() calls a "check_data" operation (if available).
If data integrity check fails, a retry to save the sent/received
data is implemented in tpm_tis_send_main()/tpm_tis_recv() functions.

Considering this commit, the following steps are done when sending a command:
   1. Host writes to TPM_STS.commandReady.
   2. Host writes command.
   3. Host checks that TPM received data is valid.
   4. If data is currupted go to step 1.

When receiving data:
   1. Host checks that TPM_STS.dataAvail is set.
   2. Host saves received data.
   3. Host checks that received data is correct.
   4. If data is currupted Host writes to TPM_STS.responseRetry and go to
      step 1.

Co-developed-by: Christophe Ricard <christophe-h.ricard@st.com>
Signed-off-by: Amir Mizinski <amirmizi6@gmail.com>
Signed-off-by: Christophe Ricard <christophe-h.ricard@st.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>