spi-nor: Remove unused file
When applying the commit 'mtd: spi-nor: Add support for w25q512jv',
spi-nor.c was accidentally re-added to the tree.

As the w25q512jv change has made it upstream we can drop the patch.
There is a subsequent change to mx66l51235f that will be addressed in a
follow up.

Signed-off-by: Joel Stanley <joel@jms.id.au>