ARM: aspeed: g5: Do not set sirq polarity
A feature was added to the aspeed vuart driver to configure the vuart
interrupt (sirq) polarity according to the LPC/eSPI strapping register.

Systems that depend on a active low behaviour (sirq_polarity set to 0)
such as OpenPower boxes also use LPC, so this relationship does not
hold.

The property was added for a Tyan S7106 system which is unfortuantly not
supported in the kernel tree. Should other systems wish to use this
feature of the driver they should add it to the machine device tree.

OpenBMC-Staging-Count: 1
Fixes: c791fc76bc72 ("arm: dts: aspeed: Add vuart aspeed,sirq-polarity-sense...")
Cc: stable@vger.kernel.org
Alexander A. Filippov <a.filippov@yadro.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>