ARM: dts: rainier: Add 'factory-reset-toggle' as GPIOF6
The state of this GPIO determines whether a factory reset has been
requested. If a physical switch is used, it can be high or low. During boot,
the software checks and records the state of this switch. If it is different
than the previous recorded state, then the read-write portions of memory are
reformatted.

Signed-off-by: Isaac Kurth <isaac.kurth@ibm.com>
Reviewed-by: Adriana Kobylak <anoo@us.ibm.com>
Link: https://lore.kernel.org/r/20210714214741.1547052-1-blisaac91@gmail.com
Signed-off-by: Joel Stanley <joel@jms.id.au>