misc: smpro-misc: Go to next stage at DIMM training failure
When there are DIMM training failure errors, boot_stage is 3 and
boot_status is 3 but current_boot_stage is 2. This makes the sysfs
return errors when the application try to access next boot stage.
The driver expects that current_boot_stage is always bigger than or
equal to boot_statge but it is not true in this case.
This commit adds workaround to make the application access the next
stage after DIMM training failure errors.

Tested:
    1. Inject DIMM training failure errors.
    2. Check DIMM training errors in REDFish/SEL log.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>