misc: smpro-misc: Go to next stage at DIMM training failure
When there are DIMM training failure errors. The boot_stage will be 3
and the boot_status is 3 but the current_boot_stage will be 2. This
cause the sysfs return the error when the application try to access
next boot stage. The driver expected that the current_boot_stage will
bigger than boot_statge but it is not true in this case. This commit
add workaround solution to support the application access the next
stage after DIMM training failure errors.

Tested:
    1. Inject DIMM training failure errors.
    2. Check DIMM training errors in REDFish/SEL log.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>