misc: smpro-errmon: Removed unnecessary stack variables initialization
These variables are never used without initialization, so stack variables
initialization are redundant and are removed in this commit.
Also switch to use the "native" int size for loop variables as it does
not provide any performance gain using smaller type but introduce potential
overflow issue when the code grow in future.

Signed-off-by: Quan Nguyen <qnguyen@amperecomputing.com>